Refactor attempt
