+++ src/runelite/constants.ts
@@ -1,11 +1,11 @@
 import {
   DIARY_VARBITS,
   DIARY_VARPS,
-} from "../services/AchievementDiaryService";
-import { COMBAT_ACHIEVEMENTS_VARPS } from "../services/CombatAchievementsService";
-import { LEAGUE_TASK_VARPS } from "../services/LeagueService";
-import { MUSIC_TRACK_VARPS } from "../services/MusicService";
-import { QUEST_VARBITS, QUEST_VARPS } from "../services/QuestService";
+} from "./transformers/AchievementDiaryService";
+import { COMBAT_ACHIEVEMENTS_VARPS } from "./transformers/CombatAchievementsService";
+import { LEAGUE_TASK_VARPS } from "./transformers/LeagueService";
+import { MUSIC_TRACK_VARPS } from "./transformers/MusicService";
+import { QUEST_VARBITS, QUEST_VARPS } from "./transformers/QuestService";
 
 export const MANIFEST_VERSION = 1;
