+++ src/runelite/data/leagueTaskVarps.json
