+++ src/runelite/enum/PlayerDataType.ts
