+++ src/runelite/enum/ProfileType.ts
