+++ src/runelite/enum/QuestCompletionState.ts
