+++ src/runelite/types/PlayerDataValue.ts
